weekly_summary/requirements.txt
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
